Refactor the gigantic ClusterListItem component: create two child ones for ClusterListItemNodeInfo and ClusterListItemReleaseInfo
